requirements/testing.txt
coverage
pytest
pytest-cov
pytest-django
mypy
django-stubs
